Move user sync helpers from plugins
