+++ cmk/gui/userdb/_user_sync.py
@@ -5,7 +5,11 @@
 
 import traceback
 from datetime import datetime
-from typing import Callable
+from typing import Callable, Literal
+
+from livestatus import SiteId
+
+from cmk.utils.site import omd_site
 
 from cmk.gui.background_job import (
     BackgroundJob,
@@ -18,11 +22,13 @@ from cmk.gui.exceptions import MKUserError
 from cmk.gui.http import request, response
 from cmk.gui.i18n import _
 from cmk.gui.log import logger as gui_logger
-from cmk.gui.plugins.userdb.utils import active_connections, user_sync_config
-from cmk.gui.site_config import is_wato_slave_site
+from cmk.gui.plugins.userdb.utils import active_connections
+from cmk.gui.site_config import get_site_config, is_wato_slave_site, site_is_local
 from cmk.gui.type_defs import Users
 from cmk.gui.userdb.store import general_userdb_job, load_users, save_users
 from cmk.gui.utils.urls import makeuri_contextless
+
+UserSyncConfig = Literal["all", "master"] | tuple[Literal["list"], list[str]] | None
 
 
 def execute_userdb_job() -> None:
@@ -58,6 +64,59 @@ def _userdb_sync_job_enabled() -> bool:
         return False
 
     return True
+
+
+def user_sync_config() -> UserSyncConfig:
+    # use global option as default for reading legacy options and on remote site
+    # for reading the value set by the Setup master site
+    default_cfg = user_sync_default_config(omd_site())
+    return get_site_config(omd_site()).get("user_sync", default_cfg)
+
+
+# Legacy option config.userdb_automatic_sync defaulted to "master".
+# Can be: None: (no sync), "all": all sites sync, "master": only master site sync
+# Take that option into account for compatibility reasons.
+# For remote sites in distributed setups, the default is to do no sync.
+def user_sync_default_config(site_name: SiteId) -> UserSyncConfig:
+    global_user_sync = _transform_userdb_automatic_sync(active_config.userdb_automatic_sync)
+    if global_user_sync == "master":
+        if site_is_local(site_name) and not is_wato_slave_site():
+            user_sync_default: UserSyncConfig = "all"
+        else:
+            user_sync_default = None
+    else:
+        user_sync_default = global_user_sync
+
+    return user_sync_default
+
+
+# Old vs:
+# ListChoice(
+#    title = _('Automatic User Synchronization'),
+#    help  = _('By default the users are synchronized automatically in several situations. '
+#              'The sync is started when opening the "Users" page in configuration and '
+#              'during each page rendering. Each connector can then specify if it wants to perform '
+#              'any actions. For example the LDAP connector will start the sync once the cached user '
+#              'information are too old.'),
+#    default_value = [ 'wato_users', 'page', 'wato_pre_activate_changes', 'wato_snapshot_pushed' ],
+#    choices       = [
+#        ('page',                      _('During regular page processing')),
+#        ('wato_users',                _('When opening the users\' configuration page')),
+#        ('wato_pre_activate_changes', _('Before activating the changed configuration')),
+#        ('wato_snapshot_pushed',      _('On a remote site, when it receives a new configuration')),
+#    ],
+#    allow_empty   = True,
+# ),
+def _transform_userdb_automatic_sync(val):
+    if val == []:
+        # legacy compat - disabled
+        return None
+
+    if isinstance(val, list) and val:
+        # legacy compat - all connections
+        return "all"
+
+    return val
 
 
 def ajax_sync() -> None:
